# docs/API/openapi.yaml
openapi: 3.0.0
info:
  title: AI Assistant API
  version: 1.0.0
  description: |
    RAG-based policy assistant API for Caribbean government staff.

    ## Authentication
    All endpoints except `/branding` require authentication via NextAuth session cookies.

    ## User Roles
    - **admin**: Full system access
    - **superuser**: Manage documents/users in assigned categories
    - **user**: Query documents in subscribed categories
  contact:
    name: AI Assistant Support
  license:
    name: Proprietary
servers:
  - url: https://ai.abhirup.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - AUTH_REQUIRED
            - ADMIN_REQUIRED
            - SUPERUSER_REQUIRED
            - NOT_FOUND
            - FORBIDDEN
            - INVALID_BODY
            - INVALID_ID
            - INVALID_DAYS
            - VALIDATION
            - VALIDATION_ERROR
            - FILE_TOO_LARGE
            - UPLOAD_LIMIT
            - INVALID_FILE_TYPE
            - DUPLICATE
            - INTERNAL
            - SERVICE_ERROR
            - LEGACY_WRITE_DISABLED
        details:
          type: string
          description: Additional context
      required:
        - error
...
    AiSetupCapabilityConfig:
      type: object
      required: [capabilityId, providerId, credentialId, modelOrServiceId, enabled]
      properties:
        capabilityId:
          type: string
          example: llm
        providerId:
          type: string
          description: An empty string in a capability update clears the capability configuration.
          example: openai
        credentialId:
          type: string
          nullable: true
          example: openai-550e8400-e29b-41d4-a716-446655440000
        modelOrServiceId:
          type: string
          nullable: true
          example: gpt-4o
        enabled:
          type: boolean
...
    AiSetupAuditEntry:
      type: object
      description: Redacted credential key-change audit entry.
      required:
        - id
        - organizationId
        - providerId
        - credentialId
        - actorUserId
        - actorEmail
        - action
        - redactedDetail
        - createdAt
      properties:
        id:
          type: integer
        organizationId:
          type: integer
          nullable: true
        providerId:
          type: string
        credentialId:
          type: string
          nullable: true
        actorUserId:
          type: integer
          nullable: true
        actorEmail:
          type: string
          format: email
          nullable: true
        action:
          type: string
          example: replaced
        redactedDetail:
          type: string
          nullable: true
          description: Redacted audit detail only; raw keys are never logged or returned.
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key for agent bot invocation
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Unauthorized"
            code: "AUTH_REQUIRED"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Admin access required"
            code: "ADMIN_REQUIRED"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Internal server error"
            code: "SERVICE_ERROR"
security:
  - sessionAuth: []
paths:
...
  /admin/ai-setup/organizations/{id}/capabilities:
    parameters:
      - name: id
        in: path
        required: true
        description: Numeric organization ID. Server-side organization scope is always enforced.
        schema:
          type: integer
    put:
      tags:
        - Admin - AI & API Setup
      summary: Save organization capability configuration
      description: |
        Upserts capability-to-provider/model/credential mappings for an organization.
        An item with an empty `providerId` clears that capability row. Provider IDs are
        validated against the server registry. Requires `super_admin` or the organization's
        server-resolved `org_admin`.
      operationId: putAiSetupCapabilities
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capabilities]
              properties:
                capabilities:
                  type: array
                  items:
                    $ref: '#/components/schemas/AiSetupCapabilityConfig'
      responses:
        '200':
          description: Capability configuration saved
          content:
            application/json:
              schema:
                type: object
                required: [ok, saved, cleared]
                properties:
                  ok:
                    type: boolean
                    example: true
                  saved:
                    type: integer
                    minimum: 0
                  cleared:
                    type: integer
                    minimum: 0
        '400':
          description: Invalid ID/body, malformed capability, or unknown provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/ai-setup/organizations/{id}/credentials/{credentialId}:
    parameters:
      - name: id
        in: path
        required: true
        description: Numeric organization ID. Server-side organization scope is always enforced.
        schema:
          type: integer
      - name: credentialId
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - Admin - AI & API Setup
      summary: Perform a credential action
      description: |
        Performs `test`, `replace`, `disable`, `enable`, or envelope-key `rotate`.
        Testing requires view access; all mutations require `super_admin` or the
        organization's server-resolved `org_admin`. Testing decrypts/resolves only on
        the server and returns availability, never the raw key. `replace` requires `secret`.
      operationId: actOnAiSetupCredential
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [test, replace, disable, enable, rotate]
                secret:
                  type: string
                  format: password
                  writeOnly: true
                  description: "Required only when `action` is `replace`; never returned."
      responses:
        '200':
          description: Action result. Test responses include provider and availability message; mutation responses include the action.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [ok, providerId, message]
                    properties:
                      ok:
                        type: boolean
                      providerId:
                        type: string
                      message:
                        type: string
                  - type: object
                    required: [ok, action]
                    properties:
                      ok:
                        type: boolean
                        example: true
                      action:
                        type: string
                        enum: [replace, disable, enable, rotate]
        '400':
          description: Invalid ID, unknown action, or missing replacement secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/ai-setup/organizations/{id}/audit:
    parameters:
      - name: id
        in: path
        required: true
        description: Numeric organization ID. Server-side organization scope is always enforced.
        schema:
          type: integer
    get:
      tags:
        - Admin - AI & API Setup
      summary: Get an organization's credential audit log
      description: |
        Returns up to 500 newest credential audit entries for the organization.
        Entries contain redacted detail only. Requires `super_admin`, or a global
        `admin` scoped to its own organization.
      operationId: getAiSetupOrganizationAudit
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Redacted organization credential audit
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries:
                    type: array
                    maxItems: 500
                    items:
                      $ref: '#/components/schemas/AiSetupAuditEntry'
        '400':
          description: Invalid organization ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...